Wrap the form in a visualizer
Multiple start modes for the visualiser:
 - Always new instance
 - if none open:               launch a new one
   else if one open:           send to that one
   else if more than one open: launch a selector

The selector is very rough and only handles double click at the moment;
we should at the very least allow the selection to happen when the user
hits `Enter`. A visual selection mode is probably also a decent idea, to
click a button (or hit a hotkey?) to allow clicking on the window to
send the data to rather than having to manually associate the Port/PID.

Use an `EventWaitHandle` to signal to the "launcher" that the new
process is up and listening. The wait handle is per process. Note,
I'm not sure what happens if you signal/set a handle and the dispose
it, so at the moment I'm holding onto it. I need to look into this.

I could probably throw the handle away after we receive data from some
connection, but that opens us up to some "rogue" process "stealing" the
process we just launched which could leave the original launcher just
sitting there depending on what happens when you actually dispose an
`EventWaitHandle`.

This is still pretty rough around the edges. I haven't tried super hard
to handle errors due to processes or sockets closing, things will
probably just violently crash.